Move field in a separate file
