+++ tests/types/pydantic/test_fields.py
@@ -24,7 +24,7 @@ def test_types(pydantic_type, field_type):
     class Model(pydantic.BaseModel):
         field: pydantic_type
 
-    @strawberry.pydantic.type(Model)
+    @strawberry.pydantic.type(Model, fields=["field"])
     class Type:
         pass
 
@@ -45,7 +45,7 @@ def test_types_optional(pydantic_type, field_type):
     class Model(pydantic.BaseModel):
         field: pydantic_type
 
-    @strawberry.pydantic.type(Model)
+    @strawberry.pydantic.type(Model, fields=["field"])
     class Type:
         pass
 
@@ -63,7 +63,7 @@ def test_conint():
     class Model(pydantic.BaseModel):
         field: pydantic.conint(lt=100)
 
-    @strawberry.pydantic.type(Model)
+    @strawberry.pydantic.type(Model, fields=["field"])
     class Type:
         pass
 
@@ -81,7 +81,7 @@ def test_constr():
     class Model(pydantic.BaseModel):
         field: pydantic.constr(max_length=100)
 
-    @strawberry.pydantic.type(Model)
+    @strawberry.pydantic.type(Model, fields=["field"])
     class Type:
         pass
 
@@ -117,6 +117,6 @@ def test_unsupported_types(pydantic_type):
 
     with pytest.raises(strawberry.pydantic.exceptions.UnsupportedTypeError):
 
-        @strawberry.pydantic.type(Model)
+        @strawberry.pydantic.type(Model, fields=["field"])
         class Type:
             pass
